refactor: Extract Device page tabs into separate components
Refactor the Device page by extracting all tab components into separate,
reusable files. This improves code organization and maintainability by
breaking down the monolithic Device.tsx file.

No functional changes, this is purely a structural refactoring.
