+++ frontend/src/pages/Device.tsx
@@ -24,7 +24,6 @@ import React, {
   useEffect,
   useState,
   useMemo,
-  useRef,
 } from "react";
 import { useParams } from "react-router-dom";
 import { ErrorBoundary } from "react-error-boundary";
@@ -36,33 +35,14 @@ import {
   useMutation,
   fetchQuery,
   useRelayEnvironment,
-  useRefetchableFragment,
-  usePaginationFragment,
 } from "react-relay/hooks";
 import type { PreloadedQuery } from "react-relay/hooks";
-import type { PayloadError, Subscription } from "relay-runtime";
-import { FormattedDate, FormattedMessage, useIntl } from "react-intl";
-import dayjs from "dayjs";
+import type { PayloadError } from "relay-runtime";
+import { FormattedMessage } from "react-intl";
 import _ from "lodash";
 
-import type { Device_PaginationQuery } from "api/__generated__/Device_PaginationQuery.graphql";
-import type { Device_batteryStatus$key } from "api/__generated__/Device_batteryStatus.graphql";
-import type { Device_hardwareInfo$key } from "api/__generated__/Device_hardwareInfo.graphql";
-import type { Device_location$key } from "api/__generated__/Device_location.graphql";
-import type { Device_baseImage$key } from "api/__generated__/Device_baseImage.graphql";
-import type { Device_osInfo$key } from "api/__generated__/Device_osInfo.graphql";
-import type { Device_runtimeInfo$key } from "api/__generated__/Device_runtimeInfo.graphql";
-import type { Device_storageUsage$key } from "api/__generated__/Device_storageUsage.graphql";
-import type { Device_systemStatus$key } from "api/__generated__/Device_systemStatus.graphql";
-import type { Device_wifiScanResults$key } from "api/__generated__/Device_wifiScanResults.graphql";
-import type { Device_networkInterfaces$key } from "api/__generated__/Device_networkInterfaces.graphql";
-import type { Device_otaOperations$key } from "api/__generated__/Device_otaOperations.graphql";
-import type { Device_cellularConnection$key } from "api/__generated__/Device_cellularConnection.graphql";
-import type { Device_deployedApplications$key } from "api/__generated__/Device_deployedApplications.graphql";
-import type { Device_deployedApplications_RefetchQuery } from "api/__generated__/Device_deployedApplications_RefetchQuery.graphql";
 import type { Device_connectionStatus$key } from "api/__generated__/Device_connectionStatus.graphql";
 import type { Device_getDevice_Query } from "api/__generated__/Device_getDevice_Query.graphql";
-import type { Device_createManualOtaOperation_Mutation } from "api/__generated__/Device_createManualOtaOperation_Mutation.graphql";
 import type { Device_updateDevice_Mutation } from "api/__generated__/Device_updateDevice_Mutation.graphql";
 import type { Device_addDeviceTags_Mutation } from "api/__generated__/Device_addDeviceTags_Mutation.graphql";
 import type { Device_removeDeviceTags_Mutation } from "api/__generated__/Device_removeDeviceTags_Mutation.graphql";
@@ -70,162 +50,37 @@ import type { Device_requestForwarderSession_Mutation } from "api/__generated__/
 import type { Device_getForwarderSession_Query } from "api/__generated__/Device_getForwarderSession_Query.graphql";
 import type { Device_getExistingDeviceTags_Query } from "api/__generated__/Device_getExistingDeviceTags_Query.graphql";
 import { Link, Route } from "Navigation";
-import AddAvailableApplications from "components/AddAvailableApplications";
 import Alert from "components/Alert";
 import Button from "components/Button";
-import CellularConnectionTabs from "components/CellularConnectionTabs";
 import Center from "components/Center";
 import ConnectionStatus from "components/ConnectionStatus";
 import Col from "components/Col";
-import DeployedApplicationsTable from "components/DeployedApplicationsTable";
 import Figure from "components/Figure";
 import Form from "components/Form";
 import LastSeen from "components/LastSeen";
 import LedBehaviorDropdown from "components/LedBehaviorDropdown";
-import Map from "components/Map";
-import OperationTable from "components/OperationTable";
 import Page from "components/Page";
 import Result from "components/Result";
 import Row from "components/Row";
 import Spinner from "components/Spinner";
 import Stack from "components/Stack";
-import StorageTable from "components/StorageTable";
-import Tabs, { Tab } from "components/Tabs";
-import WiFiScanResultsTable from "components/WiFiScanResultsTable";
-import BatteryTable from "components/BatteryTable";
-import NetworkInterfacesTable from "components/NetworkInterfacesTable";
-import BaseImageForm from "forms/BaseImageForm";
+import Tabs from "components/Tabs";
 import MultiSelect from "components/MultiSelect";
 import { FormRow as BaseFormRow, FormRowProps } from "components/FormRow";
 import assets from "assets";
-
-const DEVICE_HARDWARE_INFO_FRAGMENT = graphql`
-  fragment Device_hardwareInfo on Device {
-    capabilities
-    hardwareInfo {
-      cpuArchitecture
-      cpuModel
-      cpuModelName
-      cpuVendor
-      memoryTotalBytes
-    }
-  }
-`;
-
-const DEVICE_BASE_IMAGE_FRAGMENT = graphql`
-  fragment Device_baseImage on Device {
-    capabilities
-    baseImage {
-      name
-      version
-      buildId
-      fingerprint
-    }
-  }
-`;
-
-const DEVICE_OS_INFO_FRAGMENT = graphql`
-  fragment Device_osInfo on Device {
-    capabilities
-    osInfo {
-      name
-      version
-    }
-  }
-`;
-
-const DEVICE_LOCATION_FRAGMENT = graphql`
-  fragment Device_location on Device {
-    capabilities
-    position {
-      latitude
-      longitude
-      timestamp
-    }
-    location {
-      formattedAddress
-    }
-  }
-`;
-
-const DEVICE_STORAGE_USAGE_FRAGMENT = graphql`
-  fragment Device_storageUsage on Device {
-    capabilities
-    ...StorageTable_storageUsage
-  }
-`;
-
-const DEVICE_SYSTEM_STATUS_FRAGMENT = graphql`
-  fragment Device_systemStatus on Device {
-    capabilities
-    systemStatus {
-      memoryFreeBytes
-      taskCount
-      uptimeMilliseconds
-      timestamp
-    }
-  }
-`;
-
-const DEVICE_WIFI_SCAN_RESULTS_FRAGMENT = graphql`
-  fragment Device_wifiScanResults on Device {
-    capabilities
-    ...WiFiScanResultsTable_wifiScanResults
-  }
-`;
-
-const DEVICE_BATTERY_STATUS_FRAGMENT = graphql`
-  fragment Device_batteryStatus on Device {
-    capabilities
-    ...BatteryTable_batteryStatus
-  }
-`;
-
-const DEVICE_OTA_OPERATIONS_FRAGMENT = graphql`
-  fragment Device_otaOperations on Device
-  @refetchable(queryName: "Device_PaginationQuery") {
-    id
-    capabilities
-    otaOperations(first: $first, after: $after)
-      @connection(key: "Device_otaOperations") {
-      edges {
-        node {
-          id
-          baseImageUrl
-          status
-          createdAt
-        }
-      }
-    }
-    ...OperationTable_otaOperations
-  }
-`;
-
-const DEVICE_RUNTIME_INFO_FRAGMENT = graphql`
-  fragment Device_runtimeInfo on Device {
-    capabilities
-    runtimeInfo {
-      name
-      version
-      environment
-      url
-    }
-  }
-`;
-
-const DEVICE_CELLULAR_CONNECTION_FRAGMENT = graphql`
-  fragment Device_cellularConnection on Device {
-    capabilities
-    ...CellularConnectionTabs_cellularConnection
-  }
-`;
-
-const DEVICE_NETWORK_INTERFACES__FRAGMENT = graphql`
-  fragment Device_networkInterfaces on Device {
-    capabilities
-    ...NetworkInterfacesTable_networkInterfaces
-  }
-`;
+import DeviceHardwareInfoTab from "components/DeviceTabs/HardwareInfoTab";
+import DeviceOSInfoTab from "components/DeviceTabs/OSInfoTab";
+import DeviceRuntimeInfoTab from "components/DeviceTabs/RuntimeInfoTab";
+import DeviceBaseImageTab from "components/DeviceTabs/BaseImageTab";
+import DeviceSystemStatusTab from "components/DeviceTabs/SystemStatusTab";
+import DeviceStorageUsageTab from "components/DeviceTabs/StorageUsageTab";
+import DeviceBatteryTab from "components/DeviceTabs/BatteryTab";
+import DeviceCellularConnectionTab from "components/DeviceTabs/CellularConnectionTab";
+import DeviceNetworkInterfacesTab from "components/DeviceTabs/NetworkInterfacesTab";
+import DeviceLocationTab from "components/DeviceTabs/LocationTab";
+import DeviceWiFiScanResultsTab from "components/DeviceTabs/WiFiScanResultsTab";
+import DeviceSoftwareUpdateTab from "components/DeviceTabs/SoftwareUpdateTab";
+import DeviceApplicationsTab from "components/DeviceTabs/ApplicationsTab";
 
 const DEVICE_CONNECTION_STATUS_FRAGMENT = graphql`
   fragment Device_connectionStatus on Device {
@@ -265,74 +120,21 @@ const GET_DEVICE_QUERY = graphql`
         id
         name
       }
-      ...Device_hardwareInfo
-      ...Device_baseImage
-      ...Device_osInfo
-      ...Device_runtimeInfo
-      ...Device_location
-      ...Device_storageUsage
-      ...Device_systemStatus
-      ...Device_wifiScanResults
-      ...Device_batteryStatus
-      ...Device_otaOperations
-      ...Device_cellularConnection
-      ...Device_networkInterfaces
+      ...HardwareInfoTab_hardwareInfo
+      ...BaseImageTab_baseImage
+      ...OSInfoTab_osInfo
+      ...RuntimeInfoTab_runtimeInfo
+      ...LocationTab_location
+      ...StorageUsageTab_storageUsage
+      ...SystemStatusTab_systemStatus
+      ...WiFiScanResultsTab_wifiScanResults
+      ...BatteryTab_batteryStatus
+      ...SoftwareUpdateTab_otaOperations
+      ...CellularConnectionTab_cellularConnection
+      ...NetworkInterfacesTab_networkInterfaces
       ...Device_connectionStatus
     }
-    ...Device_deployedApplications
-  }
-`;
-
-// TODO: the fragment is defined on the RootQueryType so it can specify
-// which query to run, otherwise Relay would automatically use the `node`
-// query when refetching the fragment. However, the backend doesn't currently
-// support loading relationships (the device's deployments) through the `node`
-// query.
-// See also: https://github.com/ash-project/ash_graphql/issues/99
-const DEVICE_DEPLOYED_APPLICATIONS_FRAGMENT = graphql`
-  fragment Device_deployedApplications on RootQueryType
-  @refetchable(queryName: "Device_deployedApplications_RefetchQuery") {
-    device(id: $id) {
-      id
-      online
-      systemModel {
-        name
-      }
-      ...DeployedApplicationsTable_deployedApplications
-    }
-  }
-`;
-
-const GET_DEVICE_OTA_OPERATIONS_QUERY = graphql`
-  query Device_getDeviceOtaOperations_Query(
-    $id: ID!
-    $first: Int
-    $after: String
-  ) {
-    device(id: $id) {
-      id
-      online
-      lastConnection
-      lastDisconnection
-      ...Device_otaOperations
-    }
-  }
-`;
-
-const DEVICE_CREATE_MANUAL_OTA_OPERATION_MUTATION = graphql`
-  mutation Device_createManualOtaOperation_Mutation(
-    $input: CreateManualOtaOperationInput!
-  ) {
-    createManualOtaOperation(input: $input) {
-      result {
-        id
-        baseImageUrl
-        createdAt
-        status
-        statusCode
-        updatedAt
-      }
-    }
+    ...ApplicationsTab_deployedApplications
   }
 `;
 
@@ -454,935 +256,6 @@ const formatBytes = (bytes: number, decimals = 2) => {
   const i = Math.floor(Math.log(bytes) / Math.log(k));
 
   return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + " " + sizes[i];
-};
-
-interface DeviceHardwareInfoTabProps {
-  deviceRef: Device_hardwareInfo$key;
-}
-
-const DeviceHardwareInfoTab = ({ deviceRef }: DeviceHardwareInfoTabProps) => {
-  const intl = useIntl();
-  const { hardwareInfo, capabilities } = useFragment(
-    DEVICE_HARDWARE_INFO_FRAGMENT,
-    deviceRef,
-  );
-  if (!hardwareInfo || !capabilities.includes("HARDWARE_INFO")) {
-    return null;
-  }
-  return (
-    <Tab
-      eventKey="device-hardware-info-tab"
-      title={intl.formatMessage({
-        id: "pages.Device.hardwareInfoTab",
-        defaultMessage: "Hardware Info",
-      })}
-    >
-      <div className="mt-3">
-        <Stack gap={3}>
-          {hardwareInfo.cpuArchitecture != null && (
-            <FormRow
-              id="device-hardware-info-cpu-architecture"
-              label={
-                <FormattedMessage
-                  id="Device.hardwareInfo.cpuArchitecture"
-                  defaultMessage="CPU architecture"
-                />
-              }
-            >
-              <Form.Control
-                type="text"
-                value={hardwareInfo.cpuArchitecture}
-                readOnly
-              />
-            </FormRow>
-          )}
-          {hardwareInfo.cpuModel != null && (
-            <FormRow
-              id="device-hardware-info-cpu-model"
-              label={
-                <FormattedMessage
-                  id="Device.hardwareInfo.cpuModel"
-                  defaultMessage="CPU model code"
-                />
-              }
-            >
-              <Form.Control
-                type="text"
-                value={hardwareInfo.cpuModel}
-                readOnly
-              />
-            </FormRow>
-          )}
-          {hardwareInfo.cpuModelName != null && (
-            <FormRow
-              id="device-hardware-info-cpu-model-name"
-              label={
-                <FormattedMessage
-                  id="Device.hardwareInfo.cpuModelName"
-                  defaultMessage="CPU model name"
-                />
-              }
-            >
-              <Form.Control
-                type="text"
-                value={hardwareInfo.cpuModelName}
-                readOnly
-              />
-            </FormRow>
-          )}
-          {hardwareInfo.cpuVendor != null && (
-            <FormRow
-              id="device-hardware-info-cpu-vendor"
-              label={
-                <FormattedMessage
-                  id="Device.hardwareInfo.cpuVendor"
-                  defaultMessage="CPU vendor"
-                />
-              }
-            >
-              <Form.Control
-                type="text"
-                value={hardwareInfo.cpuVendor}
-                readOnly
-              />
-            </FormRow>
-          )}
-          {hardwareInfo.memoryTotalBytes != null && (
-            <FormRow
-              id="device-hardware-info-memory-total-bytes"
-              label={
-                <FormattedMessage
-                  id="Device.hardwareInfo.memoryTotalBytes"
-                  defaultMessage="Total memory"
-                />
-              }
-            >
-              <Form.Control
-                type="text"
-                value={formatBytes(hardwareInfo.memoryTotalBytes)}
-                readOnly
-              />
-            </FormRow>
-          )}
-        </Stack>
-      </div>
-    </Tab>
-  );
-};
-
-interface ApplicationsTabProps {
-  deviceRef: Device_deployedApplications$key;
-}
-
-const ApplicationsTab = ({ deviceRef }: ApplicationsTabProps) => {
-  const [errorFeedback, setErrorFeedback] = useState<React.ReactNode>(null);
-  const intl = useIntl();
-
-  const [{ device }, refetch] = useRefetchableFragment<
-    Device_deployedApplications_RefetchQuery,
-    Device_deployedApplications$key
-  >(DEVICE_DEPLOYED_APPLICATIONS_FRAGMENT, deviceRef);
-
-  const isOnline = useMemo(() => device?.online ?? false, [device]);
-
-  const handleRefetch = useCallback(() => {
-    refetch(
-      { id: device?.id, first: 10_000 },
-      { fetchPolicy: "store-and-network" },
-    );
-  }, [refetch, device?.id]);
-
-  useEffect(() => {
-    const intervalId = setInterval(handleRefetch, 5000);
-    return () => {
-      clearInterval(intervalId);
-    };
-  }, []);
-
-  if (!device) {
-    return null;
-  }
-
-  return (
-    <Tab
-      eventKey="applications-tab"
-      title={intl.formatMessage({
-        id: "pages.Device.ApplicationsTab",
-        defaultMessage: "Applications",
-      })}
-    >
-      <Alert
-        className="mt-3"
-        show={!!errorFeedback}
-        variant="danger"
-        onClose={() => setErrorFeedback(null)}
-        dismissible
-      >
-        {errorFeedback}
-      </Alert>
-      <div className="mt-3">
-        <h5>
-          <FormattedMessage
-            id="pages.Device.ApplicationsTab.InstallNewApp"
-            defaultMessage="Add Available Applications"
-          />
-        </h5>
-        <AddAvailableApplications
-          deviceId={device.id}
-          systemModelName={device.systemModel?.name}
-          isOnline={isOnline}
-          setErrorFeedback={setErrorFeedback}
-          onDeployComplete={handleRefetch}
-        />
-        <h5 className="mt-4">
-          <FormattedMessage
-            id="pages.Device.ApplicationsTab.DeployedApplications"
-            defaultMessage="Deployed Applications"
-          />
-        </h5>
-        <DeployedApplicationsTable
-          deviceRef={device}
-          isOnline={isOnline}
-          systemModelName={device.systemModel?.name}
-          setErrorFeedback={setErrorFeedback}
-          onDeploymentChange={handleRefetch}
-        />
-      </div>
-    </Tab>
-  );
-};
-
-interface DeviceBaseImageTabProps {
-  deviceRef: Device_baseImage$key;
-}
-
-const DeviceBaseImageTab = ({ deviceRef }: DeviceBaseImageTabProps) => {
-  const intl = useIntl();
-  const { baseImage, capabilities } = useFragment(
-    DEVICE_BASE_IMAGE_FRAGMENT,
-    deviceRef,
-  );
-  if (
-    !baseImage ||
-    Object.values(baseImage).every((value) => value === null) ||
-    !capabilities.includes("BASE_IMAGE")
-  ) {
-    return null;
-  }
-  return (
-    <Tab
-      eventKey="device-base-image-tab"
-      title={intl.formatMessage({
-        id: "pages.Device.baseImageTab",
-        defaultMessage: "Base Image",
-      })}
-    >
-      <div className="mt-3">
-        <Stack gap={3}>
-          {baseImage.name !== null && (
-            <FormRow
-              id="device-base-image-name"
-              label={
-                <FormattedMessage
-                  id="Device.baseImage.name"
-                  defaultMessage="Name"
-                />
-              }
-            >
-              <Form.Control type="text" value={baseImage.name} readOnly />
-            </FormRow>
-          )}
-          {baseImage.version !== null && (
-            <FormRow
-              id="device-base-image-version"
-              label={
-                <FormattedMessage
-                  id="Device.baseImage.version"
-                  defaultMessage="Version"
-                />
-              }
-            >
-              <Form.Control type="text" value={baseImage.version} readOnly />
-            </FormRow>
-          )}
-          {baseImage.buildId !== null && (
-            <FormRow
-              id="device-base-image-buildId"
-              label={
-                <FormattedMessage
-                  id="Device.baseImage.buildId"
-                  defaultMessage="Build identifier"
-                />
-              }
-            >
-              <Form.Control type="text" value={baseImage.buildId} readOnly />
-            </FormRow>
-          )}
-          {baseImage.fingerprint !== null && (
-            <FormRow
-              id="device-base-image-fingerprint"
-              label={
-                <FormattedMessage
-                  id="Device.baseImage.fingerprint"
-                  defaultMessage="Fingerprint"
-                />
-              }
-            >
-              <Form.Control
-                type="text"
-                value={baseImage.fingerprint}
-                readOnly
-              />
-            </FormRow>
-          )}
-        </Stack>
-      </div>
-    </Tab>
-  );
-};
-
-interface DeviceOSInfoTabProps {
-  deviceRef: Device_osInfo$key;
-}
-
-const DeviceOSInfoTab = ({ deviceRef }: DeviceOSInfoTabProps) => {
-  const intl = useIntl();
-  const { osInfo, capabilities } = useFragment(
-    DEVICE_OS_INFO_FRAGMENT,
-    deviceRef,
-  );
-  if (
-    !osInfo ||
-    Object.values(osInfo).every((value) => value === null) ||
-    !capabilities.includes("OPERATING_SYSTEM")
-  ) {
-    return null;
-  }
-  return (
-    <Tab
-      eventKey="device-os-info-tab"
-      title={intl.formatMessage({
-        id: "pages.Device.osInfoTab",
-        defaultMessage: "Operating System",
-      })}
-    >
-      <div className="mt-3">
-        <Stack gap={3}>
-          {osInfo.name !== null && (
-            <FormRow
-              id="device-os-info-name"
-              label={
-                <FormattedMessage
-                  id="Device.osInfo.name"
-                  defaultMessage="OS name"
-                />
-              }
-            >
-              <Form.Control type="text" value={osInfo.name} readOnly />
-            </FormRow>
-          )}
-          {osInfo.version !== null && (
-            <FormRow
-              id="device-os-info-version"
-              label={
-                <FormattedMessage
-                  id="Device.osInfo.version"
-                  defaultMessage="OS version"
-                />
-              }
-            >
-              <Form.Control type="text" value={osInfo.version} readOnly />
-            </FormRow>
-          )}
-        </Stack>
-      </div>
-    </Tab>
-  );
-};
-
-interface DeviceRuntimeInfoTabProps {
-  deviceRef: Device_runtimeInfo$key;
-}
-
-const DeviceRuntimeInfoTab = ({ deviceRef }: DeviceRuntimeInfoTabProps) => {
-  const intl = useIntl();
-  const { runtimeInfo, capabilities } = useFragment(
-    DEVICE_RUNTIME_INFO_FRAGMENT,
-    deviceRef,
-  );
-  if (
-    !runtimeInfo ||
-    Object.values(runtimeInfo).every((value) => value === null) ||
-    !capabilities.includes("RUNTIME_INFO")
-  ) {
-    return null;
-  }
-  return (
-    <Tab
-      eventKey="device-runtime-info-tab"
-      title={intl.formatMessage({
-        id: "pages.Device.runtimeInfoTab",
-        defaultMessage: "Runtime",
-      })}
-    >
-      <div className="mt-3">
-        <Stack gap={3}>
-          {runtimeInfo.name !== null && (
-            <FormRow
-              id="device-runtime-info-name"
-              label={
-                <FormattedMessage
-                  id="Device.runtimeInfo.name"
-                  defaultMessage="Name"
-                />
-              }
-            >
-              <Form.Control type="text" value={runtimeInfo.name} readOnly />
-            </FormRow>
-          )}
-          {runtimeInfo.version !== null && (
-            <FormRow
-              id="device-runtime-info-version"
-              label={
-                <FormattedMessage
-                  id="Device.runtimeInfo.version"
-                  defaultMessage="Version"
-                />
-              }
-            >
-              <Form.Control type="text" value={runtimeInfo.version} readOnly />
-            </FormRow>
-          )}
-          {runtimeInfo.environment !== null && (
-            <FormRow
-              id="device-runtime-info-environment"
-              label={
-                <FormattedMessage
-                  id="Device.runtimeInfo.environment"
-                  defaultMessage="Environment"
-                />
-              }
-            >
-              <Form.Control
-                type="text"
-                value={runtimeInfo.environment}
-                readOnly
-              />
-            </FormRow>
-          )}
-          {runtimeInfo.url !== null && (
-            <FormRow
-              id="device-runtime-info-url"
-              label={
-                <FormattedMessage
-                  id="Device.runtimeInfo.url"
-                  defaultMessage="URL"
-                />
-              }
-            >
-              <Form.Control type="text" value={runtimeInfo.url} readOnly />
-            </FormRow>
-          )}
-        </Stack>
-      </div>
-    </Tab>
-  );
-};
-
-interface DeviceStorageUsageTabProps {
-  deviceRef: Device_storageUsage$key;
-}
-
-const DeviceStorageUsageTab = ({ deviceRef }: DeviceStorageUsageTabProps) => {
-  const intl = useIntl();
-  const device = useFragment(DEVICE_STORAGE_USAGE_FRAGMENT, deviceRef);
-  if (!device.capabilities.includes("STORAGE")) {
-    return null;
-  }
-  return (
-    <Tab
-      eventKey="device-storage-usage-tab"
-      title={intl.formatMessage({
-        id: "pages.Device.storageTab",
-        defaultMessage: "Storage",
-      })}
-    >
-      <div className="mt-3">
-        <StorageTable deviceRef={device} />
-      </div>
-    </Tab>
-  );
-};
-
-interface DeviceLocationTabProps {
-  deviceRef: Device_location$key;
-}
-
-const DeviceLocationTab = ({ deviceRef }: DeviceLocationTabProps) => {
-  const intl = useIntl();
-  const { capabilities, position, location } = useFragment(
-    DEVICE_LOCATION_FRAGMENT,
-    deviceRef,
-  );
-  if (!position || !capabilities.includes("GEOLOCATION")) {
-    return null;
-  }
-  return (
-    <Tab
-      eventKey="device-location-tab"
-      title={intl.formatMessage({
-        id: "pages.Device.geolocationTab",
-        defaultMessage: "Geolocation",
-      })}
-    >
-      <div className="mt-3">
-        <p>
-          <FormattedMessage
-            id="pages.Device.location.lastUpdateAt"
-            defaultMessage="Last known location, updated at {date}"
-            values={{
-              date: (
-                <FormattedDate
-                  value={new Date(position.timestamp)}
-                  year="numeric"
-                  month="long"
-                  day="numeric"
-                  hour="numeric"
-                  minute="numeric"
-                />
-              ),
-            }}
-          />
-        </p>
-        <Map
-          latitude={position.latitude}
-          longitude={position.longitude}
-          popup={
-            <div>
-              {location && <p>{location.formattedAddress}</p>}
-              <p>
-                {position.latitude}, {position.longitude}
-              </p>
-            </div>
-          }
-        />
-      </div>
-    </Tab>
-  );
-};
-
-interface DeviceSystemStatusTabProps {
-  deviceRef: Device_systemStatus$key;
-}
-
-const DeviceSystemStatusTab = ({ deviceRef }: DeviceSystemStatusTabProps) => {
-  const intl = useIntl();
-  const { systemStatus, capabilities } = useFragment(
-    DEVICE_SYSTEM_STATUS_FRAGMENT,
-    deviceRef,
-  );
-  if (!systemStatus || !capabilities.includes("SYSTEM_STATUS")) {
-    return null;
-  }
-  return (
-    <Tab
-      eventKey="device-system-status-tab"
-      title={intl.formatMessage({
-        id: "pages.Device.systemStatusTab",
-        defaultMessage: "System Status",
-      })}
-    >
-      <div className="mt-3">
-        <p className="text-muted">
-          <FormattedMessage
-            id="pages.Device.systemStatus.lastUpdateAt"
-            defaultMessage="Last updated at {date}"
-            values={{
-              date: (
-                <FormattedDate
-                  value={new Date(systemStatus.timestamp)}
-                  year="numeric"
-                  month="long"
-                  day="numeric"
-                  hour="numeric"
-                  minute="numeric"
-                />
-              ),
-            }}
-          />
-        </p>
-        <Stack gap={3}>
-          {systemStatus.memoryFreeBytes != null && (
-            <FormRow
-              id="device-system-status-memory-free-bytes"
-              label={
-                <FormattedMessage
-                  id="Device.systemStatus.memoryFreeBytes"
-                  defaultMessage="Free Memory"
-                />
-              }
-            >
-              <Form.Control
-                type="text"
-                value={formatBytes(systemStatus.memoryFreeBytes)}
-                readOnly
-              />
-            </FormRow>
-          )}
-          {systemStatus.taskCount != null && (
-            <FormRow
-              id="device-system-status-task-count"
-              label={
-                <FormattedMessage
-                  id="Device.systemStatus.taskCount"
-                  defaultMessage="Active Tasks"
-                />
-              }
-            >
-              <Form.Control
-                type="text"
-                value={systemStatus.taskCount}
-                readOnly
-              />
-            </FormRow>
-          )}
-          {systemStatus.uptimeMilliseconds != null && (
-            <FormRow
-              id="device-system-status-uptime"
-              label={
-                <FormattedMessage
-                  id="Device.systemStatus.uptimeMilliseconds"
-                  defaultMessage="Last boot at"
-                />
-              }
-            >
-              <FormValue>
-                <FormattedDate
-                  value={dayjs(systemStatus.timestamp)
-                    .subtract(systemStatus.uptimeMilliseconds, "millisecond")
-                    .toDate()}
-                  year="numeric"
-                  month="long"
-                  day="numeric"
-                  hour="numeric"
-                  minute="numeric"
-                />
-              </FormValue>
-            </FormRow>
-          )}
-        </Stack>
-      </div>
-    </Tab>
-  );
-};
-
-interface DeviceWiFiScanResultsTabProps {
-  deviceRef: Device_wifiScanResults$key;
-}
-
-const DeviceWiFiScanResultsTab = ({
-  deviceRef,
-}: DeviceWiFiScanResultsTabProps) => {
-  const intl = useIntl();
-  const device = useFragment(DEVICE_WIFI_SCAN_RESULTS_FRAGMENT, deviceRef);
-  if (!device.capabilities.includes("WIFI")) {
-    return null;
-  }
-
-  return (
-    <Tab
-      eventKey="device-wifi-scan-results-tab"
-      title={intl.formatMessage({
-        id: "pages.Device.wifiScanResultsTab",
-        defaultMessage: "WiFi APs",
-      })}
-    >
-      <div className="mt-3">
-        <WiFiScanResultsTable deviceRef={device} />
-      </div>
-    </Tab>
-  );
-};
-
-interface DeviceBatteryTabProps {
-  deviceRef: Device_batteryStatus$key;
-}
-
-const DeviceBatteryTab = ({ deviceRef }: DeviceBatteryTabProps) => {
-  const intl = useIntl();
-  const device = useFragment(DEVICE_BATTERY_STATUS_FRAGMENT, deviceRef);
-  if (!device.capabilities.includes("BATTERY_STATUS")) {
-    return null;
-  }
-
-  return (
-    <Tab
-      eventKey="device-battery-tab"
-      title={intl.formatMessage({
-        id: "pages.Device.BatteryStatusTab",
-        defaultMessage: "Battery",
-      })}
-    >
-      <div className="mt-3">
-        <BatteryTable deviceRef={device} />
-      </div>
-    </Tab>
-  );
-};
-
-type SoftwareUpdateTabProps = {
-  deviceRef: Device_otaOperations$key;
-};
-
-const SoftwareUpdateTab = ({ deviceRef }: SoftwareUpdateTabProps) => {
-  const [isRefreshing, setIsRefreshing] = useState(false);
-  const [errorFeedback, setErrorFeedback] = useState<React.ReactNode>(null);
-  const intl = useIntl();
-  const relayEnvironment = useRelayEnvironment();
-
-  const { data } = usePaginationFragment<
-    Device_PaginationQuery,
-    Device_otaOperations$key
-  >(DEVICE_OTA_OPERATIONS_FRAGMENT, deviceRef);
-
-  const deviceId = data.id;
-
-  const [createOtaOperation, isCreatingOtaOperation] =
-    useMutation<Device_createManualOtaOperation_Mutation>(
-      DEVICE_CREATE_MANUAL_OTA_OPERATION_MUTATION,
-    );
-
-  const otaOperations = (
-    data.otaOperations?.edges?.map(({ node }) => node) || []
-  ).sort((a, b) => {
-    if (a.createdAt > b.createdAt) {
-      return -1;
-    }
-    if (a.createdAt < b.createdAt) {
-      return 1;
-    }
-    return 0;
-  });
-
-  const lastFinishedOperationIndex = otaOperations.findIndex(
-    ({ status }) => status === "SUCCESS" || status === "FAILURE",
-  );
-  const currentOperations =
-    lastFinishedOperationIndex === -1
-      ? otaOperations
-      : otaOperations.slice(0, lastFinishedOperationIndex);
-
-  // For now devices only support 1 update operation at a time
-  const currentOperation = currentOperations[0] || null;
-
-  // TODO: use GraphQL subscription (when available) to get updates about OTA operation
-  const subscriptionRef = useRef<Subscription | null>(null);
-  useEffect(() => {
-    return () => {
-      if (subscriptionRef.current) {
-        subscriptionRef.current.unsubscribe();
-      }
-    };
-  }, []);
-
-  useEffect(() => {
-    if (!currentOperation || isRefreshing) {
-      return;
-    }
-    const refreshTimerId = setTimeout(() => {
-      setIsRefreshing(true);
-      subscriptionRef.current = fetchQuery(
-        relayEnvironment,
-        GET_DEVICE_OTA_OPERATIONS_QUERY,
-        {
-          id: deviceId,
-          first: 10_000,
-        },
-      ).subscribe({
-        complete: () => {
-          setIsRefreshing(false);
-        },
-        error: () => {
-          setIsRefreshing(false);
-        },
-      });
-    }, 10000);
-
-    return () => {
-      clearTimeout(refreshTimerId);
-    };
-  }, [
-    currentOperation,
-    isRefreshing,
-    setIsRefreshing,
-    relayEnvironment,
-    deviceId,
-  ]);
-
-  if (!data.capabilities.includes("SOFTWARE_UPDATES")) {
-    return null;
-  }
-
-  const launchManualOTAUpdate = (file: File) => {
-    createOtaOperation({
-      variables: {
-        input: {
-          deviceId,
-          baseImageFile: file,
-        },
-      },
-      onCompleted(data, errors) {
-        if (errors) {
-          const errorFeedback = errors
-            .map(({ fields, message }) =>
-              fields.length ? `${fields.join(" ")} ${message}` : message,
-            )
-            .join(". \n");
-          return setErrorFeedback(errorFeedback);
-        }
-      },
-      onError() {
-        setErrorFeedback(
-          <FormattedMessage
-            id="pages.Device.otaUpdateCreationErrorFeedback"
-            defaultMessage="Could not start the OTA update, please try again."
-          />,
-        );
-      },
-      updater(store, data) {
-        const otaOperationId = data?.createManualOtaOperation?.result?.id;
-        if (otaOperationId) {
-          const otaOperation = store.get(otaOperationId);
-          const storedDevice = store.get(deviceId);
-          const otaOperations = storedDevice?.getLinkedRecords("otaOperations");
-          if (storedDevice && otaOperation && otaOperations) {
-            storedDevice.setLinkedRecords(
-              [otaOperation, ...otaOperations],
-              "otaOperations",
-            );
-          }
-        }
-      },
-    });
-  };
-
-  return (
-    <Tab
-      eventKey="device-software-update-tab"
-      title={intl.formatMessage({
-        id: "pages.Device.SoftwareUpdateTab",
-        defaultMessage: "Software Updates",
-      })}
-    >
-      <div className="mt-3">
-        <h5>
-          <FormattedMessage
-            id="pages.Device.SoftwareUpdateTab.manualOTAUpdate"
-            defaultMessage="Manual OTA Update"
-          />
-        </h5>
-        <Alert
-          show={!!errorFeedback}
-          variant="danger"
-          onClose={() => setErrorFeedback(null)}
-          dismissible
-        >
-          {errorFeedback}
-        </Alert>
-        <BaseImageForm
-          className="mt-3"
-          onSubmit={launchManualOTAUpdate}
-          isLoading={isCreatingOtaOperation}
-        />
-        {currentOperation && (
-          <div className="mt-3">
-            <FormattedMessage
-              id="pages.Device.SoftwareUpdateTab.updatingTo"
-              defaultMessage="Updating to image <a>{baseImageName}</a>"
-              values={{
-                a: (chunks: React.ReactNode) => (
-                  <a
-                    target="_blank"
-                    rel="noreferrer"
-                    href={currentOperation.baseImageUrl}
-                  >
-                    {chunks}
-                  </a>
-                ),
-                baseImageName: currentOperation.baseImageUrl.split("/").pop(),
-              }}
-            />
-            {isRefreshing && <Spinner size="sm" className="ms-2" />}
-          </div>
-        )}
-        <h5 className="mt-4">
-          <FormattedMessage
-            id="pages.Device.SoftwareUpdateTab.updatesHistory"
-            defaultMessage="History"
-          />
-        </h5>
-        <OperationTable deviceRef={data} />
-      </div>
-    </Tab>
-  );
-};
-
-interface DeviceCellularConnectionTabProps {
-  deviceRef: Device_cellularConnection$key;
-}
-
-const DeviceCellularConnectionTab = ({
-  deviceRef,
-}: DeviceCellularConnectionTabProps) => {
-  const intl = useIntl();
-  const device = useFragment(DEVICE_CELLULAR_CONNECTION_FRAGMENT, deviceRef);
-
-  if (!device.capabilities.includes("CELLULAR_CONNECTION")) {
-    return null;
-  }
-
-  return (
-    <Tab
-      eventKey="device-cellular-connection-tab"
-      title={intl.formatMessage({
-        id: "pages.Device.CellularConnectionTab",
-        defaultMessage: "Cellular Connection",
-      })}
-    >
-      <div className="mt-3">
-        <CellularConnectionTabs deviceRef={device} />
-      </div>
-    </Tab>
-  );
-};
-
-interface DeviceNetworkInterfacesTabProps {
-  deviceRef: Device_networkInterfaces$key;
-}
-
-const DeviceNetworkInterfacesTab = ({
-  deviceRef,
-}: DeviceNetworkInterfacesTabProps) => {
-  const intl = useIntl();
-  const device = useFragment(DEVICE_NETWORK_INTERFACES__FRAGMENT, deviceRef);
-
-  if (!device.capabilities.includes("NETWORK_INTERFACE_INFO")) {
-    return null;
-  }
-
-  return (
-    <Tab
-      eventKey="device-network-interfaces-tab"
-      title={intl.formatMessage({
-        id: "pages.Device.NetworkInterfacesTab",
-        defaultMessage: "Network Interfaces",
-      })}
-    >
-      <div className="mt-3">
-        <NetworkInterfacesTable deviceRef={device} />
-      </div>
-    </Tab>
-  );
 };
 
 interface DeviceConnectionFormRowsProps {
@@ -2020,8 +893,8 @@ const DeviceContent = ({
             <DeviceNetworkInterfacesTab deviceRef={device} />
             <DeviceLocationTab deviceRef={device} />
             <DeviceWiFiScanResultsTab deviceRef={device} />
-            <SoftwareUpdateTab deviceRef={device} />
-            <ApplicationsTab deviceRef={deviceData} />
+            <DeviceSoftwareUpdateTab deviceRef={device} />
+            <DeviceApplicationsTab deviceRef={deviceData} />
           </Tabs>
         </Stack>
       </Page.Main>
@@ -2078,5 +951,7 @@ const DevicePage = () => {
     </Suspense>
   );
 };
+
+export { FormRow, FormValue, formatBytes };
 
 export default DevicePage;
